enable running npx Doist/todoist-mcp.git (#41)
I want to run this without the inconvenience of having to build/install
first. This makes it nicer to have in an mcp config json for some
client. As long as it's not yet published with complied js sources. And
one might want to install from master or other commit'ish anyway.

You may test with
```
npx -y kirel/todoist-mcp.git#npx
```